# Init input system

+++ src/include/body.h
@@ -12,8 +12,8 @@ typedef struct{
     f16 speedY;         //speed y axis
     f16 positionX;      //position x axis  
     f16 positionY;      //position y axis 
-    u8  axisX;          //pivot point X
-	u8  axisY;          //pivot point Y
+    s16  axisX;          //pivot point X
+	s16  axisY;          //pivot point Y
     s8  direction;      // -1 or 1
     BoxCollision box;   // body box
     u8  palID;          //Character palette ID
